task: adding resume download functionality, by skipping already downloaded file

+++ services/drive.py
@@ -1,5 +1,6 @@
 from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
 import io
+import os
 
 def create_folder(service, folder_name, parent_folder_id=None):
     """Create a folder in Google Drive and return its ID."""
@@ -64,14 +65,18 @@ def copy_file(service, fileId, fileName, parent_folder_id):
 
 # To Download Files
 def downloadfiles(service, dowid, name,dfilespath):
-    request = service.files().get_media(fileId=dowid)
-    fh = io.BytesIO()
-    downloader = MediaIoBaseDownload(fh, request)
-    done = False
-    while done is False:
-        status, done = downloader.next_chunk()
-        print("Download %d%%." % int(status.progress() * 100))
-    with io.open(dfilespath + "/" + name, 'wb') as f:
-        fh.seek(0)
-        f.write(fh.read())
-
+    p = dfilespath + "/" + name
+    if not os.path.exists(p):
+      request = service.files().get_media(fileId=dowid)
+      fh = io.BytesIO()
+      downloader = MediaIoBaseDownload(fh, request)
+      done = False
+      while done is False:
+          status, done = downloader.next_chunk()
+          print("Download %d%%." % int(status.progress() * 100))
+      with io.open(dfilespath + "/" + name, 'wb') as f:
+          fh.seek(0)
+          f.write(fh.read())
+    else:
+      print("skipping because already exist: ", name)
+
